+++ .prettierignore
@@ -1,0 +1,6 @@
+# By default, ignore everything...
+/*
+
+# Except the below...
+!/public
+!/src
